update the run.py

+++ configs/run.py
@@ -6,18 +6,91 @@ import subprocess
 with open('template.py', 'r', encoding='utf-8') as file:
     template = file.read()
 
+# # 定义替换变量的值
+# test1_variables = {
+#     "num_kv_attention_head": [8, 32],
+#     # "uly_sp": [1, 2, 4, 8, 16, ],
+#     "ring_sp": [2, 4, 8, 16, 32, 64],
+#     "window_size": [1],
+#     "comm_type": ["p2p_AG"],
+#     "activation_ckpt": [False, True]
+# }
+
+# variables = test1_variables
+# root_path = "test1/"
+
 # 定义替换变量的值
-test1_variables = {
+# test2_variables = {
+#     "num_kv_attention_head": [8, 32],
+#     # "uly_sp": [1, 2, 4, 8, 16, ],
+#     "ring_sp": [64],
+#     "window_size": [2, 4, 8],
+#     "comm_type": ["double_ring"],
+#     "activation_ckpt": [False, True]
+# }
+
+# variables = test2_variables
+# root_path = "test2/"
+
+# test3_variables = {
+#     "num_kv_attention_head": [8, 32],
+#     # "uly_sp": [1, 2, 4, 8, 16, ],
+#     "ring_sp": [64],
+#     "window_size": [2, 4, 8],
+#     "comm_type": ["p2p_AG"],
+#     "activation_ckpt": [False, True]
+# }
+
+# variables = test3_variables
+# root_path = "test3/"
+
+# test4_variables = {
+#     "num_kv_attention_head": [8, 32],
+#     # "uly_sp": [1, 2, 4, 8, 16, ],
+#     "ring_sp": [16],
+#     "window_size": [2, 4, 8],
+#     "comm_type": ["double_ring"],
+#     "activation_ckpt": [False, True]
+# }
+
+# variables = test4_variables
+# root_path = "test4/"
+
+# test5_variables = {
+#     "num_kv_attention_head": [8, 32],
+#     # "uly_sp": [1, 2, 4, 8, 16, ],
+#     "ring_sp": [16],
+#     "window_size": [2, 4, 8],
+#     "comm_type": ["p2p_AG"],
+#     "activation_ckpt": [False, True]
+# }
+
+# variables = test5_variables
+# root_path = "test5/"
+
+# test6_variables = {
+#     "num_kv_attention_head": [8, 32],
+#     # "uly_sp": [1, 2, 4, 8, 16, ],
+#     "ring_sp": [8],
+#     "window_size": [2, 4, 8],
+#     "comm_type": ["double_ring"],
+#     "activation_ckpt": [False, True]
+# }
+
+# variables = test6_variables
+# root_path = "test6/"
+
+test7_variables = {
     "num_kv_attention_head": [8, 32],
     # "uly_sp": [1, 2, 4, 8, 16, ],
-    "ring_sp": [2, 4, 8, 16, 32, 64],
-    "window_size": [1],
+    "ring_sp": [8],
+    "window_size": [2, 4, 8],
     "comm_type": ["p2p_AG"],
     "activation_ckpt": [False, True]
 }
 
-variables = test1_variables
-root_path = "test1/"
+variables = test7_variables
+root_path = "test7/"
 
 
 output_folder = root_path
@@ -43,6 +116,12 @@ for i, combination in enumerate(combinations):
     # 定义输出文件名，可以使用组合索引或其他唯一标识符
     output_filename = f"kv{combination_dict['num_kv_attention_head']}_ring{combination_dict['ring_sp']}_ws{combination_dict['window_size']}_comm_{combination_dict['comm_type']}_ckpt{combination_dict['activation_ckpt']}.py"
     output_path = os.path.join(output_folder, output_filename)
+    
+    # 检查文件是否已经存在
+    if os.path.exists(output_path):
+        print(f"文件已存在，跳过生成：{output_path}")
+        continue
+    
     with open(output_path, "w", encoding='utf-8') as f:
         f.write(generated_content)
     
@@ -54,6 +133,7 @@ for i, combination in enumerate(combinations):
     log_path = os.path.join(output_folder, log_filename)
     
     # 运行命令
+    # command = f"srun -p Intern5 -N 8 -n 64 --ntasks-per-node=8 --gpus-per-task=1 python ../train.py --config ./{output_path} --profiling 2>&1 | tee '{log_path}'"
     command = f"srun -p Intern5 -N 8 -n 64 --ntasks-per-node=8 --gpus-per-task=1 python ../train.py --config ./{output_path} --profiling 2>&1 | tee '{log_path}'"
     process = subprocess.run(command, shell=True)
